requirements.txt
langserve==0.2.1
fastapi==0.111.0
langchain-community==0.2.1
sse-starlette==2.1.0
chromadb==0.5.0
python-jose==3.3.0
passlib==1.7.4
cryptography==42.0.5